+++ spec/types/uri_spec.rb
@@ -4,6 +4,10 @@ describe Attributor::URI do
 
   subject(:type) { Attributor::URI }
 
+  it 'it is not Dumpable' do
+    type.new.is_a?(Attributor::Dumpable).should_not be(true)
+  end
+
   its(:native_type) { should be ::URI::Generic }
 
   context '.example' do